chore(elements): Prep for public builds
