+++ .changeset/popular-experts-roll.md
@@ -1,0 +1,5 @@
+---
+'@clerk/elements': patch
+---
+
+Changeset to trigger `@clerk/elements` snapshot release.
