+++ .changeset/pre.json
@@ -17,7 +17,7 @@
     "@clerk/shared": "1.0.0",
     "@clerk/themes": "1.7.9",
     "@clerk/types": "3.57.0",
-    "@clerk/elements": "0.0.1"
+    "@clerk/elements": "0.0.0"
   },
   "changesets": [
     "afraid-experts-eat",
