+++ packages/elements/examples/nextjs/app/sign-in/[[...sign-in]]/page.tsx
@@ -1,16 +1,7 @@
 'use client';
 
-import {
-  GlobalError,
-  SignIn,
-  SignInContinue,
-  SignInFactorOne,
-  SignInFactorTwo,
-  SignInSocialProvider,
-  SignInStart,
-  SignInStrategy,
-  Submit,
-} from '@clerk/elements';
+import { GlobalError, Submit } from '@clerk/elements/common';
+import { Continue, Factor, SignIn, SocialProvider, Start, Verification } from '@clerk/elements/sign-in';
 
 import { H1, H2, H3, HR as Hr, P } from '@/components/design';
 import { CustomField, CustomSubmit } from '@/components/form';
@@ -21,36 +12,36 @@ export default function SignInPage() {
   return (
     <SignIn>
       <div className='m-auto w-max text-sm'>
-        <SignInStart>
+        <Start>
           <div className='flex flex-col items-center justify-center  gap-12'>
             <H1>START</H1>
 
             <GlobalError className='block text-red-400 font-mono' />
 
             <div className='flex flex-col items-stretch justify-center gap-2'>
-              <SignInSocialProvider
+              <SocialProvider
                 name='github'
                 className='flex items-center justify-center gap-4 text-white rounded bg-[#171717] px-4 py-3 text-sm shadow-sm ring-1 ring-black/[0.06] transition-all hover:bg-opacity-80'
               >
                 <SocialProviderIcon className='invert' />
                 Sign In with GitHub
-              </SignInSocialProvider>
+              </SocialProvider>
 
-              <SignInSocialProvider
+              <SocialProvider
                 name='google'
                 className='flex items-center justify-center gap-4 text-white rounded bg-[#333f61] px-4 py-3 text-sm shadow-sm ring-1 ring-black/[0.06] transition-all hover:bg-opacity-80'
               >
                 <SocialProviderIcon />
                 Sign In with Google
-              </SignInSocialProvider>
+              </SocialProvider>
 
-              <SignInSocialProvider
+              <SocialProvider
                 name='metamask'
                 className='flex items-center justify-center gap-4 text-[#161616] rounded bg-white px-4 py-3 text-sm shadow-sm ring-1 ring-black/[0.06] transition-all hover:bg-opacity-80'
               >
                 <SocialProviderIcon />
                 Sign In with Metamask
-              </SignInSocialProvider>
+              </SocialProvider>
             </div>
 
             <Hr />
@@ -71,36 +62,35 @@ export default function SignInPage() {
               <CustomSubmit>Sign In</CustomSubmit>
             </div>
           </div>
-        </SignInStart>
+        </Start>
 
-        <SignInContinue>
+        <Continue>
           <div className='flex gap-6 flex-col'>
             <H1>STRATEGIES (FIRST/SECOND FACTOR)</H1>
 
             <H2>
-              Current Factor: <SignInFactorOne>First</SignInFactorOne>
-              <SignInFactorTwo>Second</SignInFactorTwo>
+              Current Factor: <Factor first>First</Factor> | <Factor second>Second</Factor>
             </H2>
 
-            <SignInStrategy name='password'>
+            <Verification name='password'>
               <CustomField
                 label='Password'
                 name='password'
               />
 
               <CustomSubmit>Sign In</CustomSubmit>
-            </SignInStrategy>
+            </Verification>
 
-            <SignInStrategy name='email_code'>
+            <Verification name='email_code'>
               <CustomField
                 label='Email Code'
                 name='code'
               />
 
               <CustomSubmit>Sign In</CustomSubmit>
-            </SignInStrategy>
+            </Verification>
 
-            <SignInStrategy name='phone_code'>
+            <Verification name='phone_code'>
               <CustomField
                 label='Phone Code'
                 name='code'
@@ -109,9 +99,9 @@ export default function SignInPage() {
               <Submit className='px-4 py-2 b-1 bg-blue-950 bg-opacity-20 hover:bg-opacity-10 active:bg-opacity-5 rounded-md dark:bg-opacity-100 dark:hover:bg-opacity-80 dark:active:bg-opacity-50 transition'>
                 Sign In
               </Submit>
-            </SignInStrategy>
+            </Verification>
 
-            <SignInStrategy name='reset_password_email_code'>
+            <Verification name='reset_password_email_code'>
               <H3>Verify your email</H3>
 
               <P>Please check your email for a verification code...</P>
@@ -119,9 +109,9 @@ export default function SignInPage() {
               <Submit className='px-4 py-2 b-1 bg-blue-950 bg-opacity-20 hover:bg-opacity-10 active:bg-opacity-5 rounded-md dark:bg-opacity-100 dark:hover:bg-opacity-80 dark:active:bg-opacity-50 transition'>
                 Verify
               </Submit>
-            </SignInStrategy>
+            </Verification>
           </div>
-        </SignInContinue>
+        </Continue>
       </div>
 
       <SignInDebug />
