+++ packages/elements/package.json
@@ -26,6 +26,16 @@
   "author": "Clerk",
   "sideEffects": false,
   "exports": {
+    "./*": {
+      "import": {
+        "types": "./dist/react/*/index.d.mts",
+        "default": "./dist/react/*/index.mjs"
+      },
+      "require": {
+        "types": "./dist/react/*/index.d.ts",
+        "default": "./dist/react/*/index.js"
+      }
+    },
     ".": {
       "import": {
         "types": "./dist/index.d.mts",
@@ -40,7 +50,9 @@
   "main": "./dist/index.js",
   "types": "./dist/index.d.ts",
   "files": [
-    "dist"
+    "dist",
+    "!dist/metafile-*.json",
+    "!dist/**/*.map"
   ],
   "scripts": {
     "app:build": "(cd examples/nextjs && npm run build)",
@@ -49,6 +61,8 @@
     "app:e2e": "(cd examples/nextjs && npm run e2e)",
     "app:lint": "(cd examples/nextjs && npm run lint)",
     "build": "tsup",
+    "build:analyze": "tsup --metafile; open https://esbuild.github.io/analyze/",
+    "build:declarations": "tsc -p tsconfig.json",
     "clean": "rimraf ./dist",
     "dev": "tsup --watch",
     "dev:publish": "npm run dev -- --env.publish",
@@ -59,14 +73,10 @@
     "test:cache:clear": "jest --clearCache --useStderr"
   },
   "dependencies": {
-    "@clerk/clerk-react": "5.0.0-beta-v5.17",
-    "@clerk/nextjs": "5.0.0-beta-v5.20",
-    "@clerk/shared": "2.0.0-beta-v5.11",
     "@radix-ui/react-form": "^0.0.3",
     "@radix-ui/react-slot": "^1.0.2",
     "@statelyai/inspect": "^0.1.0",
     "@xstate/react": "^4.0.2",
-    "clsx": "^2.1.0",
     "xstate": "^5.4.1"
   },
   "devDependencies": {
@@ -77,15 +87,22 @@
     "eslint-config-custom": "*",
     "next": "^14.0.4",
     "tslib": "2.4.1",
+    "tsup": "^8.0.1",
     "type-fest": "^4.9.0",
     "typescript": "^5.3.3"
   },
   "peerDependencies": {
+    "@clerk/clerk-react": ">=5.0.0-beta-v5.17",
+    "@clerk/nextjs": ">=5.0.0-beta-v5.20",
+    "@clerk/shared": ">=2.0.0-beta-v5.11",
     "next": ">=13.0.4",
     "react": ">=18",
     "react-dom": ">=18"
   },
   "peerDependenciesMeta": {
+    "@clerk/nextjs": {
+      "optional": true
+    },
     "next": {
       "optional": true
     }
