+++ packages/elements/src/react/common/index.ts
@@ -1,0 +1,26 @@
+// TODO: Split and invert the naming
+
+export {
+  Field as ClerkField,
+  FieldError as ClerkFieldError,
+  FieldState as ClerkFieldState,
+  Form as ClerkForm,
+  GlobalError as ClerkGlobalError,
+  Input as ClerkInput,
+  Label as ClerkLabel,
+  Submit as ClerkSubmit,
+  //
+  Field,
+  FieldError,
+  FieldState,
+  Form,
+  GlobalError,
+  Input,
+  Label,
+  Submit,
+} from '~/react/common/form';
+
+export {
+  SocialProviderIcon,
+  SocialProviderIcon as ClerkSocialProviderIcon,
+} from '~/react/common/third-party-providers/social-provider';
