+++ packages/elements/src/react/sign-in/index.ts
@@ -1,0 +1,13 @@
+'use client';
+
+export { SignInContinue, SignInContinue as Continue } from './continue';
+export { SignInRoot, SignInRoot as SignIn, SignInRoot as Root } from './root';
+export {
+  SignInSocialProvider as SocialProvider,
+  SignInSocialProviderIcon as SocialProviderIcon,
+} from './social-providers';
+export { SignInStart as Start } from './start';
+export { SignInFactor as Factor, SignInVerification as Verification } from './verifications';
+
+// TODO: Move contexts from /internals to /react
+export { useSignInFlow, useSignInFlowSelector } from '~/internals/machines/sign-in/sign-in.context';
