+++ packages/elements/src/react/sign-up/index.ts
@@ -1,0 +1,13 @@
+'use client';
+
+export { SignUpContinue as Continue } from './continue';
+export { SignUpRoot, SignUpRoot as SignUp, SignUpRoot as Root } from './root';
+export {
+  SignUpSocialProvider as SocialProvider,
+  SignUpSocialProviderIcon as SocialProviderIcon,
+} from './social-providers';
+export { SignUpStart as Start } from './start';
+export { SignUpVerification as Verification, SignUpVerify as Verify } from './verifications';
+
+// TODO: Move contexts from /internals to /react
+export { useSignUpFlow, useSignUpFlowSelector } from '~/internals/machines/sign-up/sign-up.context';
